backend/requirements/base.txt
fastapi
pydantic_settings
sqlalchemy
greenlet
asyncpg
psycopg2
alembic
pyjwt
bcrypt